Food_recommend/requirements.txt
Django>=3.2,<5.0
Pillow>=8.0.0
django-crispy-forms>=1.14.0
django-cleanup>=6.0.0 